backend/requirements.txt
fastapi
uvicorn
python-multipart
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform
pydantic
python-dotenv
pypdf
